pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multiplayer-engine"
version = "1.0.0"
description = "Real-Time Multiplayer Game Networking Engine — a CN course project"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "pygame-ce>=2.5.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "cryptography>=46.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
multiplayer-engine-server = "server.server:main"
multiplayer-engine-client = "client.client:main"
multiplayer-engine-analyze = "analysis.plot_results:main"

[tool.setuptools.packages.find]
include = ["common*", "server*", "client*", "analysis*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
